+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed playback halting on unplayable files such as MIDI by skipping to the next track ([#386])
 
 ## [1.8.1] - 2026-02-14
 ### Changed
@@ -125,6 +127,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#269]: https://github.com/FossifyOrg/Music-Player/issues/269
 [#298]: https://github.com/FossifyOrg/Music-Player/issues/298
 [#361]: https://github.com/FossifyOrg/Music-Player/issues/361
+[#386]: https://github.com/FossifyOrg/Music-Player/issues/386
 
 [Unreleased]: https://github.com/FossifyOrg/Music-Player/compare/1.8.1...HEAD
 [1.8.1]: https://github.com/FossifyOrg/Music-Player/compare/1.8.0...1.8.1